Implement basic concept of shell statements.
The shell operates on a concept of statements now rather than a blob of
raw text. Raw input is converted to a statement structure that can be
processed further. This should allow for 'JIT' variable substitution to
take place prior to execution of the statement. It will also lay the
foundation for scripts which will just be a chain of statements.